+++ test/subdir_ffi/src/project_ffi.erl
@@ -1,0 +1,6 @@
+-module(project_ffi).
+
+-export([log/1]).
+
+log(Message) ->
+  erlang:display(Message).